Loki: Rename public/app/plugins/datasource/loki/query_utils.ts file to follow convention
