+++ public/app/plugins/datasource/loki/components/AnnotationsQueryEditor.tsx
@@ -5,7 +5,7 @@ import { AnnotationQuery } from '@grafana/data';
 import { EditorField, EditorRow, Input } from '@grafana/ui';
 
 // Types
-import { getNormalizedLokiQuery } from '../query_utils';
+import { getNormalizedLokiQuery } from '../queryUtils';
 import { LokiQuery, LokiQueryType } from '../types';
 
 import { LokiOptionFields } from './LokiOptionFields';
